% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavitopen.R
\name{cavitopen}
\alias{cavitopen}
\title{CAVITOPEN - an R implementation of Herve Cochard et al.'s PLC simulator}
\usage{
cavitopen(wps = seq(0, -12, by = -1), rotor_length = 27, p50_true = -7.38,
  s50_true = 51.28, pcav_open = -0.5, vessel_length_mean = 10,
  vessel_length_max = 10, vessel_dist = c("exp", "linear"),
  d_stepsize = 1e-04)
}
\arguments{
\item{wps}{Water potentials (MPa) at which to evaluate the various PLCs.}

\item{rotor_length}{Length of the rotor used (cm)}

\item{p50_true}{True P50 for non-open vessels}

\item{s50_true}{True slope at P50 for non-open vessels}

\item{pcav_open}{Threshold water potential for cavitation of open vessels}

\item{vessel_length_mean}{Mean vessel length, used if \code{vessel_dist} is linear.}

\item{vessel_length_max}{Maximum vessel length, used if \code{vessel_dist} is exponential.}

\item{d_stepsize}{Step size along segment length to integrate PLC. Probably don't ever have to change this.}
}
\description{
Description of the method to be provided. Based on Excel spreadsheet 'Cavitopen 3.xlsx'.
}
\examples{

# The default settings reproduce the example in 'cavitopen 3.xlsx'.
run1 <- cavitopen()
plot(run1)

# Smoother curves
run2 <- cavitopen(wps = seq(0, -12, length=101))
plot(run2, type='l')
}
